# Benchmarks: multi-byte switching overhead vs. mb_* performance