… tests/benchmarks/class-multibyte-switching-bench.php