Benchmark results from PHP 7.2 with opcache. Relative PHP 5.6 results are similar, though everything takes a bit longer.

```13 subjects, 156 iterations, 15,600 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 67.925 [18,934.579 18,934.732] 76.860 (μs)
⅀T: 2,953,794.250μs μSD/r 161.367μs μRSD/r: 2.429%
suite: 133c9d19c6bb492015223463715d427d99ba5fe6, date: 2017-12-17, stime: 11:31:39
+---------------------------+-----------------------------+--------+----------------------------------------------------+------+-----+------------+---------+---------+---------+---------+-------+--------+-----------+
| benchmark                 | subject                     | groups | params                                             | revs | its | mem_peak   | best    | mean    | mode    | worst   | stdev | rstdev | diff      |
+---------------------------+-----------------------------+--------+----------------------------------------------------+------+-----+------------+---------+---------+---------+---------+-------+--------+-----------+
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"long":"ASCII"}                                   | 200  | 2   | 961,192b   | 3.6ms   | 3.6ms   | 3.6ms   | 3.6ms   | 0.0ms | 0.07%  | 48.34x    |
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,192b   | 3.5ms   | 3.5ms   | 3.5ms   | 3.6ms   | 0.0ms | 0.69%  | 47.49x    |
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,192b   | 0.3ms   | 0.3ms   | 0.3ms   | 0.3ms   | 0.0ms | 5.05%  | 4.17x     |
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,192b   | 0.3ms   | 0.3ms   | 0.3ms   | 0.3ms   | 0.0ms | 4.93%  | 4.11x     |
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 5.91%  | 1.06x     |
| Multibyte_Switching_Bench | bench_mb_strlen             |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 1.96%  | 1.16x     |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"long":"ASCII"}                                   | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 1.13%  | 1.85x     |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,216b   | 3.6ms   | 3.7ms   | 3.7ms   | 3.7ms   | 0.0ms | 1.04%  | 49.10x    |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 8.71%  | 1.08x     |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 0.3ms   | 0.3ms   | 0.3ms   | 0.3ms   | 0.0ms | 3.49%  | 3.90x     |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 8.76%  | 1.00x     |
| Multibyte_Switching_Bench | bench_switched_strlen       |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 8.47%  | 1.12x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"long":"ASCII"}                                   | 200  | 2   | 961,192b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 4.52%  | 1.31x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,192b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 0.98%  | 1.22x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,192b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 5.60%  | 1.17x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,192b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 7.89%  | 1.17x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 1.88%  | 1.19x     |
| Multibyte_Switching_Bench | bench_mb_substr             |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 6.61%  | 1.17x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"long":"ASCII"}                                   | 200  | 2   | 961,216b   | 0.2ms   | 0.2ms   | 0.2ms   | 0.2ms   | 0.0ms | 2.96%  | 2.09x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 4.70%  | 1.80x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 1.95%  | 1.11x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 0.56%  | 1.31x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 3.03%  | 1.00x     |
| Multibyte_Switching_Bench | bench_switched_substr       |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 7.16%  | 1.33x     |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"long":"ASCII"}                                   | 200  | 2   | 1,501,576b | 121.2ms | 121.8ms | 121.8ms | 122.4ms | 0.6ms | 0.51%  | 1,636.02x |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"long":"UTF-8"}                                   | 200  | 2   | 1,500,392b | 120.6ms | 120.7ms | 120.7ms | 120.9ms | 0.2ms | 0.15%  | 1,621.90x |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 7.7ms   | 7.9ms   | 7.9ms   | 8.2ms   | 0.2ms | 2.95%  | 106.41x   |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 7.5ms   | 7.6ms   | 7.6ms   | 7.7ms   | 0.1ms | 1.36%  | 101.72x   |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.4ms   | 0.5ms   | 0.5ms   | 0.5ms   | 0.0ms | 0.52%  | 6.05x     |
| Multibyte_Switching_Bench | bench_mb_str_split          |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.4ms   | 0.5ms   | 0.5ms   | 0.5ms   | 0.0ms | 3.59%  | 6.15x     |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"long":"ASCII"}                                   | 200  | 2   | 1,500,856b | 120.3ms | 120.7ms | 120.7ms | 121.1ms | 0.4ms | 0.34%  | 1,621.88x |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"long":"UTF-8"}                                   | 200  | 2   | 1,499,672b | 120.5ms | 120.6ms | 120.6ms | 120.7ms | 0.1ms | 0.10%  | 1,620.02x |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 7.6ms   | 7.6ms   | 7.6ms   | 7.7ms   | 0.0ms | 0.55%  | 102.52x   |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 7.5ms   | 7.6ms   | 7.6ms   | 7.7ms   | 0.1ms | 0.77%  | 102.10x   |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.3ms   | 0.3ms   | 0.3ms   | 0.4ms   | 0.0ms | 4.27%  | 4.61x     |
| Multibyte_Switching_Bench | bench_preg_split            |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.4ms   | 0.4ms   | 0.4ms   | 0.4ms   | 0.0ms | 1.82%  | 4.91x     |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"long":"ASCII"}                                   | 200  | 2   | 1,501,584b | 26.8ms  | 26.9ms  | 26.9ms  | 27.0ms  | 0.1ms | 0.26%  | 361.13x   |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"long":"UTF-8"}                                   | 200  | 2   | 1,500,416b | 121.0ms | 121.9ms | 121.9ms | 122.9ms | 1.0ms | 0.79%  | 1,637.91x |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,240b   | 1.7ms   | 1.7ms   | 1.7ms   | 1.7ms   | 0.0ms | 0.18%  | 22.60x    |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,240b   | 7.5ms   | 7.6ms   | 7.6ms   | 7.7ms   | 0.1ms | 1.39%  | 102.66x   |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,264b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 4.96%  | 1.80x     |
| Multibyte_Switching_Bench | bench_switched_str_split    |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,264b   | 0.5ms   | 0.5ms   | 0.5ms   | 0.5ms   | 0.0ms | 3.27%  | 6.32x     |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"long":"ASCII"}                                   | 200  | 2   | 961,216b   | 83.5ms  | 83.8ms  | 83.8ms  | 84.1ms  | 0.3ms | 0.40%  | 1,125.88x |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,216b   | 83.5ms  | 84.0ms  | 84.0ms  | 84.5ms  | 0.5ms | 0.60%  | 1,128.19x |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 5.4ms   | 5.5ms   | 5.5ms   | 5.7ms   | 0.1ms | 2.28%  | 74.54x    |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 5.5ms   | 5.6ms   | 5.6ms   | 5.6ms   | 0.0ms | 0.67%  | 74.98x    |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.4ms   | 0.4ms   | 0.4ms   | 0.5ms   | 0.0ms | 7.46%  | 5.90x     |
| Multibyte_Switching_Bench | bench_mb_strtolower         |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.4ms   | 0.5ms   | 0.5ms   | 0.5ms   | 0.0ms | 3.97%  | 6.06x     |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"long":"ASCII"}                                   | 200  | 2   | 961,240b   | 4.3ms   | 4.4ms   | 4.4ms   | 4.4ms   | 0.0ms | 1.08%  | 58.67x    |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,240b   | 83.1ms  | 83.5ms  | 83.5ms  | 84.0ms  | 0.5ms | 0.54%  | 1,121.75x |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,240b   | 0.3ms   | 0.4ms   | 0.4ms   | 0.4ms   | 0.0ms | 2.45%  | 4.76x     |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,240b   | 5.6ms   | 5.6ms   | 5.6ms   | 5.6ms   | 0.0ms | 0.39%  | 75.55x    |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,264b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 0.26%  | 1.10x     |
| Multibyte_Switching_Bench | bench_switched_strtolower   |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,264b   | 0.5ms   | 0.5ms   | 0.5ms   | 0.5ms   | 0.0ms | 2.76%  | 6.30x     |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"long":"ASCII"}                                   | 200  | 2   | 961,216b   | 145.4ms | 145.8ms | 145.8ms | 146.1ms | 0.4ms | 0.24%  | 1,958.20x |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,216b   | 137.3ms | 141.2ms | 141.2ms | 145.1ms | 3.9ms | 2.76%  | 1,896.66x |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 8.9ms   | 9.0ms   | 9.0ms   | 9.0ms   | 0.0ms | 0.49%  | 120.29x   |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 8.8ms   | 8.8ms   | 8.8ms   | 8.9ms   | 0.1ms | 0.61%  | 118.28x   |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.5ms   | 0.6ms   | 0.6ms   | 0.6ms   | 0.0ms | 0.58%  | 7.39x     |
| Multibyte_Switching_Bench | bench_mb_strtoupper         |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.5ms   | 0.6ms   | 0.6ms   | 0.6ms   | 0.0ms | 2.43%  | 7.49x     |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"long":"ASCII"}                                   | 200  | 2   | 961,240b   | 5.1ms   | 5.1ms   | 5.1ms   | 5.1ms   | 0.0ms | 0.59%  | 68.33x    |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,240b   | 140.3ms | 143.2ms | 143.2ms | 146.0ms | 2.9ms | 1.99%  | 1,923.43x |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,240b   | 0.4ms   | 0.4ms   | 0.4ms   | 0.4ms   | 0.0ms | 1.85%  | 5.68x     |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,240b   | 10.0ms  | 10.0ms  | 10.0ms  | 10.0ms  | 0.0ms | 0.34%  | 134.24x   |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,264b   | 0.1ms   | 0.1ms   | 0.1ms   | 0.1ms   | 0.0ms | 3.20%  | 1.33x     |
| Multibyte_Switching_Bench | bench_switched_strtoupper   |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,264b   | 0.6ms   | 0.6ms   | 0.6ms   | 0.6ms   | 0.0ms | 1.07%  | 8.30x     |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"long":"ASCII"}                                   | 200  | 2   | 961,216b   | 12.2ms  | 12.3ms  | 12.3ms  | 12.3ms  | 0.1ms | 0.54%  | 164.89x   |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,216b   | 11.9ms  | 12.3ms  | 12.3ms  | 12.6ms  | 0.4ms | 3.00%  | 164.63x   |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,216b   | 0.8ms   | 0.8ms   | 0.8ms   | 0.8ms   | 0.0ms | 3.69%  | 10.73x    |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,216b   | 0.8ms   | 0.8ms   | 0.8ms   | 0.8ms   | 0.0ms | 1.72%  | 10.59x    |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,240b   | 0.2ms   | 0.2ms   | 0.2ms   | 0.2ms   | 0.0ms | 6.86%  | 2.80x     |
| Multibyte_Switching_Bench | bench_preg_replace          |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,240b   | 0.2ms   | 0.2ms   | 0.2ms   | 0.2ms   | 0.0ms | 3.55%  | 2.91x     |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"long":"ASCII"}                                   | 200  | 2   | 961,240b   | 8.7ms   | 8.7ms   | 8.7ms   | 8.7ms   | 0.0ms | 0.07%  | 116.74x   |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"long":"UTF-8"}                                   | 200  | 2   | 961,240b   | 12.1ms  | 12.1ms  | 12.1ms  | 12.1ms  | 0.0ms | 0.08%  | 162.02x   |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"medium":"ASCII"}                                 | 200  | 2   | 961,240b   | 0.5ms   | 0.6ms   | 0.6ms   | 0.6ms   | 0.0ms | 1.65%  | 7.41x     |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"medium":"UTF-8"}                                 | 200  | 2   | 961,240b   | 0.7ms   | 0.8ms   | 0.8ms   | 0.8ms   | 0.0ms | 0.83%  | 10.11x    |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"html":"aaabbbcccdddeeefffggghhh"}                | 200  | 2   | 961,264b   | 0.2ms   | 0.2ms   | 0.2ms   | 0.2ms   | 0.0ms | 1.68%  | 2.95x     |
| Multibyte_Switching_Bench | bench_switched_preg_replace |        | {"html":"\u00e4\u00e4\u00e4bbbcccdddeeefffggghhh"} | 200  | 2   | 961,264b   | 0.2ms   | 0.2ms   | 0.2ms   | 0.2ms   | 0.0ms | 1.00%  | 2.74x     |
+---------------------------+-----------------------------+--------+----------------------------------------------------+------+-----+------------+---------+---------+---------+---------+-------+--------+-----------+```