@mundschenk-at 2017-12-20
```
3 subjects, 60 iterations, 600 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 3.000 [9.653 9.311] 4.370 (μs)
⅀T: 579.150μs μSD/r 0.620μs μRSD/r: 7.030%
suite: 133c9d450ca0e69965aa131d8c579dfbf196edb5, date: 2017-12-20, stime: 21:48:46
+-----------------+--------------------+--------+---------------------------------------------------------+------+-----+----------+----------+----------+----------+----------+---------+--------+-------+
| benchmark       | subject            | groups | params                                                  | revs | its | mem_peak | best     | mean     | mode     | worst    | stdev   | rstdev | diff  |
+-----------------+--------------------+--------+---------------------------------------------------------+------+-----+----------+----------+----------+----------+----------+---------+--------+-------+
| Functions_Bench | bench_uchr         |        | [8205,8204,768,769,771,772,775,776,784,803,805,814,817] | 100  | 10  | 961,432b | 16.600μs | 17.598μs | 16.983μs | 19.520μs | 1.005μs | 5.71%  | 5.08x |
| Functions_Bench | bench_uchr         |        | [8205]                                                  | 100  | 10  | 961,112b | 6.810μs  | 7.136μs  | 6.997μs  | 7.720μs  | 0.280μs | 3.92%  | 2.06x |
| Functions_Bench | bench_json_uchr    |        | [8205,8204,768,769,771,772,775,776,784,803,805,814,817] | 100  | 10  | 961,432b | 12.510μs | 13.253μs | 12.726μs | 14.720μs | 0.797μs | 6.01%  | 3.83x |
| Functions_Bench | bench_json_uchr    |        | [8205]                                                  | 100  | 10  | 961,112b | 3.990μs  | 4.338μs  | 4.303μs  | 4.860μs  | 0.236μs | 5.44%  | 1.25x |
| Functions_Bench | bench_intlchar_chr |        | [8205,8204,768,769,771,772,775,776,784,803,805,814,817] | 100  | 10  | 961,456b | 11.270μs | 12.126μs | 11.607μs | 14.140μs | 0.940μs | 7.75%  | 3.50x |
| Functions_Bench | bench_intlchar_chr |        | [8205]                                                  | 100  | 10  | 961,136b | 3.000μs  | 3.464μs  | 3.249μs  | 4.370μs  | 0.462μs | 13.34% | 1.00x |
+-----------------+--------------------+--------+---------------------------------------------------------+------+-----+----------+----------+----------+----------+----------+---------+--------+-------+
```